+++ src/main/java/com/techelevator/Candy.java
@@ -1,10 +1,11 @@
 package com.techelevator;
 
 import java.io.PrintWriter;
+import java.math.BigDecimal;
 
 public class Candy extends Product {
 
-    public Candy(String name, double price){
+    public Candy(String name, BigDecimal price){
         super(name, price);
     }
